+++ packages/cli/test/unit/controllers/executions.controller.test.ts
@@ -21,15 +21,11 @@ describe('ExecutionsController', () => {
 	});
 
 	describe('getMany', () => {
+		const NO_EXECUTIONS = { count: 0, estimated: false, results: [] };
+
 		it('if no status provided, should look for all active plus latest 20 finished executions', async () => {
 			workflowSharingService.getSharedWorkflowIds.mockResolvedValue(['123']);
-			const allActiveAndLatestTwentyFinishedSpy = jest
-				.spyOn(executionsController, 'allActiveAndLatestTwentyFinished')
-				.mockResolvedValue({
-					results: [],
-					count: 0,
-					estimated: false,
-				});
+			executionService.findAllActiveAndLatestTwentyFinished.mockResolvedValue(NO_EXECUTIONS);
 
 			const req = mock<ExecutionRequest.GetMany>({
 				rangeQuery: { kind: 'range', workflowId: undefined, status: undefined },
@@ -37,19 +33,13 @@ describe('ExecutionsController', () => {
 
 			await executionsController.getMany(req);
 
-			expect(allActiveAndLatestTwentyFinishedSpy).toHaveBeenCalled();
+			expect(executionService.findAllActiveAndLatestTwentyFinished).toHaveBeenCalled();
 			expect(executionService.findRangeWithCount).not.toHaveBeenCalled();
 		});
 
 		it('if status provided as empty array, should look for all active plus latest 20 finished executions', async () => {
 			workflowSharingService.getSharedWorkflowIds.mockResolvedValue(['123']);
-			const allActiveAndLatestTwentyFinishedSpy = jest
-				.spyOn(executionsController, 'allActiveAndLatestTwentyFinished')
-				.mockResolvedValue({
-					results: [],
-					count: 0,
-					estimated: false,
-				});
+			executionService.findAllActiveAndLatestTwentyFinished.mockResolvedValue(NO_EXECUTIONS);
 
 			const req = mock<ExecutionRequest.GetMany>({
 				rangeQuery: { kind: 'range', workflowId: undefined, status: [] },
@@ -57,19 +47,13 @@ describe('ExecutionsController', () => {
 
 			await executionsController.getMany(req);
 
-			expect(allActiveAndLatestTwentyFinishedSpy).toHaveBeenCalled();
+			expect(executionService.findAllActiveAndLatestTwentyFinished).toHaveBeenCalled();
 			expect(executionService.findRangeWithCount).not.toHaveBeenCalled();
 		});
 
 		it('if status provided, should look for a range of executions based on the query', async () => {
 			workflowSharingService.getSharedWorkflowIds.mockResolvedValue(['123']);
-			const allActiveAndLatestTwentyFinishedSpy = jest
-				.spyOn(executionsController, 'allActiveAndLatestTwentyFinished')
-				.mockResolvedValue({
-					results: [],
-					count: 0,
-					estimated: false,
-				});
+			executionService.findAllActiveAndLatestTwentyFinished.mockResolvedValue(NO_EXECUTIONS);
 
 			const req = mock<ExecutionRequest.GetMany>({
 				rangeQuery: { kind: 'range', workflowId: undefined, status: ['success'] },
@@ -77,7 +61,7 @@ describe('ExecutionsController', () => {
 
 			await executionsController.getMany(req);
 
-			expect(allActiveAndLatestTwentyFinishedSpy).not.toHaveBeenCalled();
+			expect(executionService.findAllActiveAndLatestTwentyFinished).not.toHaveBeenCalled();
 			expect(executionService.findRangeWithCount).toHaveBeenCalledWith(req.rangeQuery);
 		});
 	});
